commit pas bien

+++ src/main/java/com/gestion/demo/repository/simple/SexeRepo.java
@@ -3,6 +3,6 @@ package com.gestion.demo.repository.simple;
 import org.springframework.data.jpa.repository.JpaRepository;
 import com.gestion.demo.model.Simple.Sexe;
 
-public interface SexeRepo extends JpaRepository<Sexe, Long> {
+public interface SexeRepo extends JpaRepository<Sexe, Integer> {
 
 }
